Address Redis persistence review feedback
