+++ .changeset/calm-redis-clear.md
@@ -1,0 +1,5 @@
+---
+"effect": patch
+---
+
+Ensure clearing an empty Redis-backed persistence store succeeds.
